Tweak floating point calculations
- For clarity, use `(float)` conversion in benchmark calculations with integers.
(Thanks to @Pharap).

+++ Test0_Baseline.ino
@@ -11,7 +11,7 @@ void testBaseline(void) {
     Serial.print(F_CPU,DEC);                    // 16000000 (expected)
     logConsole(text::hertz);                    // " Hz\n"
     logConsole(text::oneCycle);                 // "1 cycle             : "
-    Serial.print((1000000.0/(float)F_CPU),4);   // 0.0625 (625ns expected)
+    Serial.print(((float)1000000/(float)F_CPU),4);   // 0.0625 (625ns expected)
     logConsole(text::micros);                   // " µs\n"
 
     benchStartTime = millis();
@@ -47,11 +47,11 @@ void testBaseline(void) {
     }
     benchEndTime = millis();
 
-    benchResult = ( (float)benchEndTime - (float)benchStartTime ) / (100.0 * 10000.0); // Calculate benchmark time for an individual loop
-    benchResult = benchResult * 1000.0; // Convert milli (10^-3) to micro (10^-6) seconds
-    benchOverhead = benchResult - (20.0 * (1000000.0/(float)F_CPU)); // Calculate overhead of nested FOR loops by subtracting 20x instruction cycles in microseconds
+    benchResult = ( (float)benchEndTime - (float)benchStartTime ) / ((float)100 * (float)10000); // Calculate benchmark time for an individual loop
+    benchResult = benchResult * (float)1000; // Convert milli (10^-3) to micro (10^-6) seconds
+    benchOverhead = benchResult - ((float)20 * ((float)1000000/(float)F_CPU)); // Calculate overhead of nested FOR loops by subtracting 20x instruction cycles in microseconds
     benchResult = benchResult - benchOverhead; // Remove overhead from results
-    benchResult = benchResult / 20.0; // Determine the result for an individual instruction
+    benchResult = benchResult / (float)20; // Determine the result for an individual instruction
 
     logConsole(text::nop);                  // "nop                 : "
     Serial.print(benchResult,4);            // ~0.0625 (625ns expected)
@@ -67,8 +67,8 @@ void testBaseline(void) {
     }
     benchEndTime = millis();
 
-    benchResult = ( (float)benchEndTime - (float)benchStartTime ) / (100.0 * 10000.0); // Calculate benchmark time for an individual loop
-    benchResult = benchResult * 1000.0; // Convert milli (10^-3) to micro (10^-6) seconds
+    benchResult = ( (float)benchEndTime - (float)benchStartTime ) / ((float)100 * (float)10000); // Calculate benchmark time for an individual loop
+    benchResult = benchResult * (float)1000; // Convert milli (10^-3) to micro (10^-6) seconds
     benchResult = benchResult - benchOverhead; // Remove loop overhead from results
     benchOverhead += benchResult; // Include the time for incrementing the loop counter
