+++ Test1_Monopixel.ino
@@ -22,7 +22,7 @@ void testMonopixel() {
 
 
     // --- Sprites::drawOverwrite ---
-    benchAverage = 0.0;
+    benchAverage = (float)0;
     for (i = 0; i < 3; i = i + 1) {
         loopCount = 0;
         benchStartTime = millis();
@@ -71,7 +71,7 @@ void testMonopixel() {
 
 
     // --- Sprites::drawErase ---
-    benchAverage = 0.0;
+    benchAverage = (float)0;
     arduboy.fillScreen(WHITE); // Subtract the inverted sprite from 'white' screen
     arduboy.invert(true); // Invert the display for visual consistency
     arduboy.display();
@@ -95,14 +95,14 @@ void testMonopixel() {
     arduboy.fillScreen(BLACK); // Back to black to avoid white flash
     arduboy.invert(false); // Restore display
     arduboy.display();
-    benchAverage /= (4.0 * 3.0); // 4 function calls x 3 loops
+    benchAverage /= ((float)4 * (float)3); // 4 function calls x 3 loops
     logConsole(text::spritesErase);             // "Sprites Erase       : "
     Serial.print(benchAverage,3);               // Result (3 d.p.)
     logConsole(text::micros, newLine);          // " µs\n"
     
     
     // --- Sprites::drawSelfMasked ---
-    benchAverage = 0.0;
+    benchAverage = (float)0;
     for (i = 0; i < 3; i = i + 1) {
         loopCount = 0;
         benchStartTime = millis();
@@ -119,14 +119,14 @@ void testMonopixel() {
         benchmark(loopCount); // Benchmark for given loops
         benchAverage = benchAverage + benchResult;
     }
-    benchAverage /= (4.0 * 3.0); // 4 function calls x 3 loops
+    benchAverage /= ((float)4 * (float)3); // 4 function calls x 3 loops
     logConsole(text::spritesSelfMask);          // "Sprites Self masked : "
     Serial.print(benchAverage,3);               // Result (3 d.p.)
     logConsole(text::micros);                   // " µs\n"
     
     
     // --- arduboy.drawBitmap ~ Sprites::drawSelfMasked ---
-    benchAverage = 0.0;
+    benchAverage = (float)0;
     for (i = 0; i < 3; i = i + 1) {
         loopCount = 0;
         benchStartTime = millis();
@@ -143,14 +143,14 @@ void testMonopixel() {
         benchmark(loopCount); // Benchmark for given loops
         benchAverage = benchAverage + benchResult; // Update the running average
     }
-    benchAverage /= (4.0 * 3.0); // 4 function calls x 3 loops
+    benchAverage /= ((float)4 * (float)3); // 4 function calls x 3 loops
     logConsole(text::bitmapsStd);               // "Bitmaps Standard    : "
     Serial.print(benchAverage,3);               // Result (3 d.p.)
     logConsole(text::micros);                   // " µs\n"
     
     
     // --- arduboy.drawSlowXYBitmap ~  Sprites::drawSelfMasked ---
-    benchAverage = 0.0;
+    benchAverage = (float)0;
     for (i = 0; i < 3; i = i + 1) {
         loopCount = 0;
         benchStartTime = millis();
@@ -167,14 +167,14 @@ void testMonopixel() {
         benchmark(loopCount); // Benchmark for given loops
         benchAverage = benchAverage + benchResult; // Update the running average
     }
-    benchAverage /= (4.0 * 3.0); // 4 function calls x 3 loops
+    benchAverage /= ((float)4 * (float)3); // 4 function calls x 3 loops
     logConsole(text::bitmapsSlow);              // "Bitmaps Slow XY     : "
     Serial.print(benchAverage,3);               // Result (3 d.p.)
     logConsole(text::micros);                   // " µs\n"
     
     
     // --- arduboy.drawCompressed ~  Sprites::drawSelfMasked ---
-    benchAverage = 0.0;
+    benchAverage = (float)0;
     for (i = 0; i < 3; i = i + 1) {
         loopCount = 0;
         benchStartTime = millis();
@@ -191,7 +191,7 @@ void testMonopixel() {
         benchmark(loopCount); // Benchmark for given loops
         benchAverage = benchAverage + benchResult; // Update the running average
     }
-    benchAverage /= (4.0 * 3.0); // 4 function calls x 3 loops
+    benchAverage /= ((float)4 * (float)3); // 4 function calls x 3 loops
     logConsole(text::bitmapsComp);              // "Bitmaps Compressed  : "
     Serial.print(benchAverage,3);               // Result (3 d.p.)
     logConsole(text::micros, newLine);          // " µs\n"
@@ -200,7 +200,7 @@ void testMonopixel() {
     
     
     // --- Sprites::drawExternalMask ---
-    benchAverage = 0.0;
+    benchAverage = (float)0;
     for (i = 0; i < 3; i = i + 1) {
         loopCount = 0;
         benchStartTime = millis();
@@ -217,14 +217,14 @@ void testMonopixel() {
         benchmark(loopCount); // Benchmark for given loops
         benchAverage = benchAverage + benchResult;
     }
-    benchAverage /= (4.0 * 3.0); // 4 function calls x 3 loops
+    benchAverage /= ((float)4 * (float)3); // 4 function calls x 3 loops
     logConsole(text::spritesExtMask);           // "Sprites Ext' mask   : "
     Serial.print(benchAverage,3);               // Result (3 d.p.)
     logConsole(text::micros);                   // " µs\n"
     
     
     // --- Sprites::drawPlusMask ---
-    benchAverage = 0.0;
+    benchAverage = (float)0;
     for (i = 0; i < 3; i = i + 1) {
         loopCount = 0;
         benchStartTime = millis();
@@ -241,14 +241,14 @@ void testMonopixel() {
         benchmark(loopCount); // Benchmark for given loops
         benchAverage = benchAverage + benchResult; // Update the running average
     }
-    benchAverage /= (4.0 * 3.0); // 4 function calls x 3 loops
+    benchAverage /= ((float)4 * (float)3); // 4 function calls x 3 loops
     logConsole(text::spritesPlusMask);          // "Sprites Plus mask   : "
     Serial.print(benchAverage,3);               // Result (3 d.p.)
     logConsole(text::micros);                   // " µs\n"
     
     
     // --- Sprites::drawSelfMasked with extra masking step ---
-    benchAverage = 0.0;
+    benchAverage = (float)0;
     for (i = 0; i < 3; i = i + 1) {
         loopCount = 0;
         benchStartTime = millis();
@@ -269,14 +269,14 @@ void testMonopixel() {
         benchmark(loopCount); // Benchmark for given loops
         benchAverage = benchAverage + benchResult; // Update the running average
     }
-    benchAverage /= (4.0 * 3.0); // 4 function calls x 3 loops
+    benchAverage /= ((float)4 * (float)3); // 4 function calls x 3 loops
     logConsole(text::spritesEraseMasked);       // "Sprites Self w/mask : "
     Serial.print(benchAverage,3);               // Result (3 d.p.)
     logConsole(text::micros);                   // " µs\n"
     
     
     // --- arduboy.drawBitmap with extra masking step ---
-    benchAverage = 0.0;
+    benchAverage = (float)0;
     for (i = 0; i < 3; i = i + 1) {
         loopCount = 0;
         benchStartTime = millis();
@@ -297,14 +297,14 @@ void testMonopixel() {
         benchmark(loopCount); // Benchmark for given loops
         benchAverage = benchAverage + benchResult; // Update the running average
     }
-    benchAverage /= (4.0 * 3.0); // 4 function calls x 3 loops
+    benchAverage /= ((float)4 * (float)3); // 4 function calls x 3 loops
     logConsole(text::bitmapsStdMasked);         // "Bitmaps Std w/mask  : "
     Serial.print(benchAverage,3);               // Result (3 d.p.)
     logConsole(text::micros);                   // " µs\n"
     
     
     // --- arduboy.drawSlowXYBitmap with extra masking step ---
-    benchAverage = 0.0;
+    benchAverage = (float)0;
     for (i = 0; i < 3; i = i + 1) {
         loopCount = 0;
         benchStartTime = millis();
@@ -325,14 +325,14 @@ void testMonopixel() {
         benchmark(loopCount); // Benchmark for given loops
         benchAverage = benchAverage + benchResult; // Update the running average
     }
-    benchAverage /= (4.0 * 3.0); // 4 function calls x 3 loops
+    benchAverage /= ((float)4 * (float)3); // 4 function calls x 3 loops
     logConsole(text::bitmapsSlowMasked);        // "Bitmaps Slow w/mask : "
     Serial.print(benchAverage,3);               // Result (3 d.p.)
     logConsole(text::micros);                   // " µs\n"
     
     
     // --- arduboy.drawCompressed with extra masking step ---
-    benchAverage = 0.0;
+    benchAverage = (float)0;
     for (i = 0; i < 3; i = i + 1) {
         loopCount = 0;
         benchStartTime = millis();
@@ -353,7 +353,7 @@ void testMonopixel() {
         benchmark(loopCount); // Benchmark for given loops
         benchAverage = benchAverage + benchResult; // Update the running average
     }
-    benchAverage /= (4.0 * 3.0); // 4 function calls x 3 loops
+    benchAverage /= ((float)4 * (float)3); // 4 function calls x 3 loops
     logConsole(text::bitmapsCompMasked);        // "Bitmaps Comp w/mask : "
     Serial.print(benchAverage,3);               // Result (3 d.p.)
     logConsole(text::micros);                   // " µs\n"
